[Improve] Rename `mmcls.data` to `mmcls.structures`.
