+++ mmcls/utils/setup_env.py
@@ -73,11 +73,11 @@ def register_all_modules(init_default_scope: bool = True) -> None:
             https://github.com/open-mmlab/mmengine/blob/main/docs/en/tutorials/registry.md
             Defaults to True.
     """  # noqa
-    import mmcls.data  # noqa: F401,F403
     import mmcls.datasets  # noqa: F401,F403
     import mmcls.engine  # noqa: F401,F403
     import mmcls.evaluation  # noqa: F401,F403
     import mmcls.models  # noqa: F401,F403
+    import mmcls.structures  # noqa: F401,F403
     import mmcls.visualization  # noqa: F401,F403
 
     if not init_default_scope:
