+++ tests/test_evaluation/test_metrics/test_multi_label.py
@@ -6,8 +6,8 @@ import sklearn.metrics
 import torch
 from mmengine.evaluator import Evaluator
 
-from mmcls.data import ClsDataSample
 from mmcls.evaluation.metrics import AveragePrecision, MultiLabelMetric
+from mmcls.structures import ClsDataSample
 from mmcls.utils import register_all_modules
 
 register_all_modules()
