+++ tests/test_models/test_classifiers.py
@@ -5,9 +5,9 @@ from unittest.mock import MagicMock
 import torch
 from mmengine import ConfigDict
 
-from mmcls.data import ClsDataSample
 from mmcls.models import ImageClassifier
 from mmcls.registry import MODELS
+from mmcls.structures import ClsDataSample
 from mmcls.utils import register_all_modules
 
 register_all_modules()
